+++ python/dalex/dalex/dataset_level/_aggregated_profiles/checks.py
@@ -1,4 +1,20 @@
 import numpy as np
+import pandas as pd
+
+
+def check_variables(variables):
+    # treating variables as list simplifies code
+    if variables is not None and not isinstance(variables, (str, list, np.ndarray, pd.Series)):
+        raise TypeError("variables must be None or str or numpy.ndarray od pandas.Series")
+
+    if variables is None:
+        variables_ = None
+    elif isinstance(variables, str):
+        variables_ = [variables]
+    else:
+        variables_ = list(variables)
+
+    return variables_
 
 
 def check_variable_type(variable_type):
@@ -56,8 +72,15 @@ def create_x(all_profiles, variable_type):
 
 
 def check_groups(groups):
-    if groups is not None and not isinstance(groups, str):
-        raise TypeError("groups must be None or str")
-
     # treating groups as list simplifies code
-    return [] if groups is None else [groups]
+    if groups is not None and not isinstance(groups, (str, list, np.ndarray, pd.Series)):
+        raise TypeError("groups must be None or str or numpy.ndarray od pandas.Series")
+
+    if groups is None:
+        groups_ = []
+    elif isinstance(groups, str):
+        groups_ = [groups]
+    else:
+        groups_ = list(groups)
+
+    return groups_