+++ python/dalex/NEWS.md
@@ -1,5 +1,8 @@
 dalex (development)
 ----------------------------------------------------------------
+*feature:* parameter `N` in `model_profile` can be set to `None`, to select all observations 
+* *fix:* `label` and `groups` fixed in `model_profile`
+* *input:* `groups` and `variable` parameters in `model_profile` can be: `str`, `list`, `numpy.ndarray`, `pandas.Series`
 * *fix:* `check_label` returned only a first letter
 * *bugfix:* removed the conversion of `all_variables` to `str` in
 `prepare_all_variables`, which caused an error in `model_profile` ([#214](https://github.com/ModelOriented/DALEX/issues/214))